Improve deeplink handling
