+++ src/lib/sync-connect-site-deeplink.ts
@@ -1,0 +1,16 @@
+import { sendIpcEventToRenderer } from 'src/ipc-utils';
+
+/**
+ * Handles the sync-connect-site deeplink callback.
+ * This function is called when a user initiates a sync connection from WordPress.com
+ * and is redirected back to the app via wpcom-local-dev://sync-connect-site
+ */
+export async function handleSyncConnectSiteDeeplink( urlObject: URL ): Promise< void > {
+	const { searchParams } = urlObject;
+	const remoteSiteId = parseInt( searchParams.get( 'remoteSiteId' ) ?? '' );
+	const studioSiteId = searchParams.get( 'studioSiteId' );
+
+	if ( remoteSiteId && studioSiteId ) {
+		void sendIpcEventToRenderer( 'sync-connect-site', { remoteSiteId, studioSiteId } );
+	}
+}
